Fix error caused by components list
