+++ src/renderer/components/RouterComponents/Description/Description.scss
@@ -1,0 +1,7 @@
+.hello {
+    color: red;
+}
+
+.description{
+    color: blue;
+}
