+++ src/renderer/constants/ComponentsList.ts
@@ -1,0 +1,6 @@
+export const COMPONENTS_LIST = {
+    COMPONENT1: 'component1',
+    COMPONENT2: 'component2',
+    COMPONENT3: 'component3',
+    DESCRIPTION: 'description'
+  }
